# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Api endpoints for exposing data models.",
    "version" : "v1",
    "title" : "Model Catalog API",
    "license" : {
      "name" : "Apache License Version 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "localhost",
  "basePath" : "/",
...
  "paths" : {
...
    "/api/v1/models/{modelId}/artifacts/{artifactId}" : {
      "get" : {
        "tags" : [ "artifacts-controller" ],
        "summary" : "Returns artifact metadata",
        "description" : "Privilege level: Consumer of this endpoint must have a valid access token",
        "operationId" : "retrieveArtifactUsingGET",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Model id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "artifactId",
          "in" : "path",
          "description" : "Artifact id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "SUCCESS",
            "schema" : {
              "$ref" : "#/definitions/ArtifactDTO"
            }
          },
          "500" : {
            "description" : "Internal server error, e.g. error getting artifact metadata"
          },
          "404" : {
            "description" : "Model or artifact not Found"
          }
        }
      },
      "delete" : {
        "tags" : [ "artifacts-controller" ],
        "summary" : "Deletes artifact",
        "description" : "Privilege level: Consumer of this endpoint must have a valid access token",
        "operationId" : "deleteArtifactUsingDELETE",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Model id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "artifactId",
          "in" : "path",
          "description" : "Artifact id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Deleted",
            "schema" : {
              "$ref" : "#/definitions/ArtifactDTO"
            }
          },
          "500" : {
            "description" : "Internal server error, e.g. error saving model metadata"
          },
          "404" : {
            "description" : "Model or artifact not Found"
          }
        }
      }
    },
    "/api/v1/models/{modelId}/artifacts/{artifactId}/file" : {
      "get" : {
        "tags" : [ "artifacts-controller" ],
        "summary" : "Downloads artifact file",
        "description" : "Privilege level: Consumer of this endpoint must have a valid access token",
        "operationId" : "retrieveArtifactFileUsingGET",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/octet-stream", "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "modelId",
          "in" : "path",
          "description" : "Model id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "artifactId",
          "in" : "path",
          "description" : "Artifact id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "SUCCESS",
            "schema" : {
              "$ref" : "#/definitions/Resource"
            }
          },
          "500" : {
            "description" : "Internal server error, e.g. error getting artifact file"
          },
          "404" : {
            "description" : "Model or artifact not Found"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "ArtifactDTO" : {
      "type" : "object",
      "properties" : {
        "actions" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "filename" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "location" : {
          "type" : "string"
        }
      }
    },
    "InputStream" : {
      "type" : "object"
    },
    "Resource" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "file" : {
          "type" : "file"
        },
        "filename" : {
          "type" : "string"
        },
        "inputStream" : {
          "$ref" : "#/definitions/InputStream"
        },
        "open" : {
          "type" : "boolean"
        },
        "readable" : {
          "type" : "boolean"
        },
        "uri" : {
          "$ref" : "#/definitions/URI"
        },
        "url" : {
          "$ref" : "#/definitions/URL"
        }
      }
    },
...
    "URI" : {
      "type" : "object",
      "properties" : {
        "absolute" : {
          "type" : "boolean"
        },
        "authority" : {
          "type" : "string"
        },
        "fragment" : {
          "type" : "string"
        },
        "host" : {
          "type" : "string"
        },
        "opaque" : {
          "type" : "boolean"
        },
        "path" : {
          "type" : "string"
        },
        "port" : {
          "type" : "integer",
          "format" : "int32"
        },
        "query" : {
          "type" : "string"
        },
        "rawAuthority" : {
          "type" : "string"
        },
        "rawFragment" : {
          "type" : "string"
        },
        "rawPath" : {
          "type" : "string"
        },
        "rawQuery" : {
          "type" : "string"
        },
        "rawSchemeSpecificPart" : {
          "type" : "string"
        },
        "rawUserInfo" : {
          "type" : "string"
        },
        "scheme" : {
          "type" : "string"
        },
        "schemeSpecificPart" : {
          "type" : "string"
        },
        "userInfo" : {
          "type" : "string"
        }
      }
    },
    "URL" : {
      "type" : "object",
      "properties" : {
        "authority" : {
          "type" : "string"
        },
        "content" : {
          "type" : "object"
        },
        "defaultPort" : {
          "type" : "integer",
          "format" : "int32"
        },
        "file" : {
          "type" : "string"
        },
        "host" : {
          "type" : "string"
        },
        "path" : {
          "type" : "string"
        },
        "port" : {
          "type" : "integer",
          "format" : "int32"
        },
        "protocol" : {
          "type" : "string"
        },
        "query" : {
          "type" : "string"
        },
        "ref" : {
          "type" : "string"
        },
        "userInfo" : {
          "type" : "string"
        }
      }
    }
  }
}